reorgnize tests folder for built-in operators
